Cleans up analytics_integration module
