+++ scanners/analytics_integration/README.md
