--- scanners/scanner_analytics_integration/Dockerfile.rhel7
@@ -1,16 +1,0 @@
-FROM registry.centos.org/centos/centos:latest
-
-LABEL INSTALL='docker run --rm --privileged -v /etc/atomic.d/:/host/etc/atomic.d/ $IMAGE sh /install.sh'
-
-# fix for employee subscription
-RUN yum -y install deltarpm yum-utils --disablerepo=*-eus-* --disablerepo=*-htb-* \
-    --disablerepo=*-ha-* --disablerepo=*-rt-* --disablerepo=*-lb-* --disablerepo=*-rs-* --disablerepo=*-sap-*
-RUN yum-config-manager --disable *-eus-* *-htb-* *-ha-* *-rt-* *-lb-* *-rs-* *-sap-* > /dev/null
-
-# Install python-docker-py to spin up container using scan script
-RUN yum -y update && \
-    yum-config-manager -q -y --enable rhel-7-server-extras-rpm && \
-    yum -y install atomic python-docker-py && \
-    yum clean all
-
-ADD analytics-integration integration.py install.sh osio_token.txt /
