--- scanners/scanner_analytics_integration/analytics-integration
@@ -1,17 +1,0 @@
-type: scanner
-scanner_name: analytics-integration
-image_name: scanner-analytics-integration:rhel7
-default_scan: register
-custom_args: ["-v", "/var/run/docker.sock:/var/run/docker.sock", "-e", "IMAGE_NAME=$IMAGE_NAME", "-e", "SERVER=$SERVER", "-e", "GITSHA=$GITSHA", "-e", "GITURL=$GITURL"]
-scans: [
-  {
-    name: register,
-    args: ["python", "integration.py", "register"],
-    description: "Registers image under test at analytics server for scanning."
-  },
-  {
-    name: report,
-    args: ["python", "integration.py", "report"],
-    description: "Gets the report for image under test from server."
-  }
-]
