--- scanners/scanner_analytics_integration/cccp.yml
@@ -1,1 +1,0 @@
-job-id: scanner-analytics-integration
