--- scanners/scanner_analytics_integration/install.sh
@@ -1,4 +1,0 @@
-# /bin/sh
-
-echo "Installing configuration for for analytics-integration scanner"
-cp -v /analytics-integration /host/etc/atomic.d
