--- scanners/scanner_analytics_integration/osio_token.txt
@@ -1,1 +1,0 @@
-please_fill_your_actual_token_here
